Избавился от дублирования. Сделал считывалку html-файлов
