+++ test/common.js
@@ -2,21 +2,11 @@
 
 'use strict';
 
-var fs = require('fs');
-var html = fs.readFileSync('index.html', 'utf-8');
 require('should');
 
-var tags = [];
+var tags = require('./getDifferentTags');
 
-html.match(/\<\s*[a-z]+/ig).forEach(function(tag){
-    tag = tag.replace('<', '');
-
-    if (tags.indexOf(tag) === -1) {
-        tags.push(tag);
-    }
-});
-
-describe('Запрещенные теги', function(){
+describe('Запрещенные теги.', function(){
     it('Не должно быть тега <i>', function () {
         (tags.indexOf('i') === -1).should.be.eql(true);
     });
